api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.20
reportlab==4.2.5
cryptography==44.0.0
PyJWT==2.10.1
passlib[argon2]==1.7.4
email-validator==2.2.0
